add new cruds

+++ app/routers/users.py
@@ -1,6 +1,6 @@
 from http import HTTPStatus
-from typing import Iterable, Type
 from fastapi import APIRouter, HTTPException
+from fastapi_pagination import Page
 from app.database import users
 from app.models.User import User, UserCreate, UserUpdate
 
@@ -19,18 +19,17 @@ def get_user(user_id: int) -> User:
 
 
 @router.get("/", status_code=HTTPStatus.OK)
-def get_users() -> Iterable[User]:
+def get_users() -> Page[User]:
     return users.get_users()
 
 
 @router.post("/", status_code=HTTPStatus.CREATED)
-def create_user(user: User) -> User:
-    UserCreate.model_validate(user.model_dump())
+def create_user(user: UserCreate) -> User:
     return users.create_user(user)
 
 
 @router.patch("/{user_id}", status_code=HTTPStatus.OK)
-def update_user(user_id: int, user: User) -> Type[User]:
+def update_user(user_id: int, user: User) -> User:
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
     UserUpdate.model_validate(user.model_dump())
@@ -41,5 +40,8 @@ def update_user(user_id: int, user: User) -> Type[User]:
 def delete_user(user_id: int):
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
+    user = users.get_user(user_id)
+    if not user:
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
     users.delete_user(user_id)
     return {"message": "User deleted"}
